TreeListItems: Loading Projects from MobX Store
